+++ packages/plonk/src/timers/Frames.ts
@@ -1,4 +1,4 @@
-import ms, { TimeFormat, type FPS } from '../math/ms';
+import ms, { TimeFormat, type FPS } from '../utils/ms';
 import Metro, { type TimerCallback, type MetroOptions } from './Metro';
 
 export type FramesOptions = {
