+++ packages/plonk/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as now } from './now';
+export { default as ms, type FPS, TimeFormat } from './ms';
